refactor: rename unit.ts to units.ts
Solves issue #22 point 3
