+++ src/common/index.ts
@@ -5,4 +5,4 @@ export * from './date-fields'
 export * from './esday-inst'
 export * from './is'
 export * from './str-utils'
-export * from './unit'
+export * from './units'
